Split apart Analyser into separate files to simplity reasoning
